requirements.txt
asgiref==3.7.2
#backports.zoneinfo==0.2.1
Django==3.0.8
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
gunicorn==21.2.0
#mysqlclient==2.2.0
packaging==23.1
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
Pygments==2.16.1